requirements.txt
pandas==2.2.0
selenium==4.17.2
parsel==1.8.1
requests==2.31.0
argparse==1.4.0
chromedriver-binary-auto